Create FoldSettingRule
This allows to easily change the fold setting in tests.
The fold setting defines the behavior of whether the screen stays on or
off when folding the device.

Also refactors SystemSettingRule to use a shell command to write the
setting, as there is a security exception when trying to write the
setting directly using the Settings class.

Bug: 288245610
Test: atest PlatformScenarioTests:FoldToLauncher
Change-Id: Id9123101e7b5f1797aec19a3e0125a0b67234af9