Wait for fold + sleep animations to finish before completing fold action
There's a new feature that doesn't immediately go to screen off
after the device is folded. We want the device to consistently start
from the same state (folded with screen off) to measure latency.
However, the `uiDevice.sleep()` isn't sufficient. Although the device
is not interactive (screenOn = false), there are still animations
that need to finish to get into an idle folded with screen off state.

Therefore, we add a delay/wait after the "sleep" request so that the device is truly in a stable, folded screen off state before proceeding
to its next action.

Fixes: 319996759
Flag: NONE
Test: UnfoldToLockscreenMicrobenchmark#testUnfold
Change-Id: I1f4da27f27aa108f9be3b64b2f866df6dd1255f5