Extract `BitmapDiffer` and `ScreenshotAsserterFactory` interfaces out of the `ScreenshotTestRule`.
The `ScreenshotTestRule` combines many responsibilities (capturing screenshots via `ScreenshotAsserter`, diffing screenshots) in one class. Since the class is wrapped in outer *ScreenshotTestRules, which are concerned with setting up the environment, reuse-ability is limited.

This change works around that by exposing distinct interfaces for creating [ScreenshotAsserter] instances or just diffing images, without relying exposing the rule itself.

This is in preparation for the `MotionTestRule`, should be able to rely on the same environment setup as the `*ScreenshotTestRule`s without having to duplicate the same setup code.

Test: Refactoring Test infrastructure - covered by existing tests
Bug: 322324387
Change-Id: I3b571d44d5f9b684a546b5de689bfc031547bc51