Factor a [DiffResultExportStrategy] out of the [ScreenshotTestRule]
Separates the logic to write and announce screenshot diffs. This will
allow to export debug screenshots to Scuba for motion tests, without
requiring an actual golden image to be checked in.

This CL intentionally focuses on extracting the existing code, without
cleaning up the logic that is now in [ScubaExportStrategy] and
[DevicelessDevMachineExportStrategy]. Follow-up CLs will clean the
`reportResult` signature. There is room to radically simplify the code
in [DevicelessDevMachineExportStrategy], but that is beyond this CLs
scope.

See go/android-animation-tests-design

Test: Existing unit tests
Test: ScubaExportStrategyTest
Bug: 322324387
Change-Id: Ie684a613bcce158633971d9e15a04dcccd8b8abe