Fold [GoldenImagePathManager] into [GoldenPathManager] #cleanup
This unblocks the use of the extension-aware functions introduced in
[GoldenPathManager].

If this breaks the build, please check whether a simple refactor of
GoldenImagePathManager to GoldenPathManager solves the issue first;
there might be a race condition between this cleanup and the
introduction of new GoldenImagePathManager usages. If so, fixing forward
is might be preferable over reverting/resubmitting this across all these
reporsitories.

Bug: 322324387
Test: Existing unit tests
Change-Id: I1ecc7b0289a349c55b313b895564f64ef8c88d3e