UInputStylus: send updateTimeBase event before replaying
The uinput command's timing of event injections recently changed [0], so
that when a delay command is sent it will cause the following inject
command(s) to be scheduled relative to the end of the last delay, even
if that is a time has passed. Since stylus tests wait for various things
(which can take multiple seconds) before replaying the recorded events,
the injections from the recording were being scheduled relative to
events that had happened before the long wait. This led to many events
being scheduled for times that had already passed, and therefore a bunch
were injected at once, overflowing the kernel's buffers and causing
events to be dropped.

To fix this, use the new updateTimeBase command just before replaying
events, to tell uinput that the injections we're about to schedule
should be scheduled relative to the current time.

[0]: change Ieaa4f2f06d5e0b7d13abc3afa474948996db7c0a

Bug: 330844071
Test: check for SYN_DROPPED events in the output of:
      $ adb shell getevent -lt
      while running the inking test:
      $ atest 'PlatformScenarioTests:android.platform.test.scenario.sysui.stylus.StylusInkingTest#writeText_appearsInShowcaseApp'
Change-Id: I4fedbd6bbd3b99b669507b15307b3f30141bace6