Rename iterations on the method name, rather than the class name.
Also pass the annotations from the original test description down to the
renamed description.

This change fixes the issue that AndroidJUnitRunner's annotation
filtering does not work with renamed-iteration microbenchmarks, because
it can look for annotations on the test class. With the old renaming
scheme, the test class becomes NULL; this change enables the test class
to be kept.

Bug: 331696130
Test: Unit tests updated.
Change-Id: Ia69c2a1b26c376b653bd136523f521232ff1b645