Golden update change auto-generated from Scuba: invocation I66700010258890370
Updating to match what is generated in post-submit. This should finally
get the test green, so that I can then switch over the tests to be
included in presubmit, and catch issues before submitting

Investigating where the black bar comes from will be done once the test
runs in presubmit.

Bug: 322324387
Test: Fixes existing unit tests
Change-Id: I0fdbfbdc7a86fe1f0a7501d597ee49693680e1e6