Remove Kotlin files excluded in platform_testing/ ktfmt hook
This CL removes the excluded files list from the ktfmt hook so that we
can migrate to using the built-in ktfmt hook (that does not support
excluding files).

This list of excluded files was introduced at the time because we mostly
wanted to avoid merge conflicts between the different *dev* branches we
had to juggle with, but now that we all work in main I think that it's
fine to remove and even good for consistency accross our codebase.

See b/230545175#comment4 and b/230545175#comment5 for details.

Bug: 342153000
Test: N/A
Flag: NONE Repohook Update
Change-Id: I835a3324803fcebe9780ab2f26aeed428b60574a