Add @NoMetricRule to Microbenchmark and Functional runner
Adds @NoMetricRule similar to @NoMetricBefore/After.

The order is still different between the runners,
but it allows to at least enforce test rule vs @NoMetricBefore/After
methods ordering.

The ordering is not matched as it requires
copying/reimplementing a bunch of classes that are
package-private in JUnit runner, but technically
should be possible.

Functional runner order:
- NoMetricRule before
- NoMetricBefore
- Rule before
- Before
- Test body
...

Microbenchmark runner order:
- NoMetricRule before
- Rule before
- NoMetricBefore
- Before
- Test body
...

Test: atest MicrobenchmarkTest
Test: atest FunctionalTest
Bug: 353672921
Flag: none
Change-Id: Ib0b0bdd54f7b5af16420b02db1f7ee454c129da4