Add headroom mode option to memhog preparer.
The headroom mode continuously allocates easily reclaimable memory as
one way to measure memory headroom on device. The memory headroom can be
measured as the memory use of the memhog progress when it reaches an
equilibrium point where the rate of memhog allocation matches the rate
of memhog pages reclaimed by the system under that level of memory
pressure.

See go/mpts-headroom-test for more info.

Test: abdt runs of MPTS in both "hog" and "headroom" modes.
Flag: TEST_ONLY

Change-Id: Ib8c59e1744d13c47773262a27858a65513ea8737