Ignore SurfaceViewBackground layers in VisibleLayersShownMoreThanOneConsecutiveEntry assertion
This assertion checks that all layers that are visible on the trace, are visible for at least 2 consecutive entries. The background for a SurfaceView might precede the SurfaceView showing by one frame, meaning
that the background disappears in the next frame when the SurfaceView shows up.

Bug: 304067015
Test: flicker
Flag: EXEMPT it's flicker test infra
Change-Id: Ia434ce7b9e314ac897f68e6db86db6db049dad08